shims-303 mapid toInt

+++ spark-extension-shims-spark303/src/main/scala/org/apache/spark/sql/execution/blaze/plan/ArrowShuffleExchangeExec.scala
@@ -147,7 +147,7 @@ case class ArrowShuffleExchangeExec(
           .nativeShuffleWrite(
             rdd.asInstanceOf[MapPartitionsRDD[_, _]].prev.asInstanceOf[NativeRDD],
             dep,
-            mapId,
+            mapId.toInt,
             context,
             partition)
       }
